+++ cmd/rr/cmd/serve.go
@@ -36,12 +36,15 @@ func init() {
 		RunE:  serveHandler,
 	})
 
-	signal.Notify(stopSignal, syscall.SIGTERM)
-	signal.Notify(stopSignal, syscall.SIGINT)
+	signal.Notify(stopSignal, os.Interrupt, os.Kill, syscall.SIGTERM)
 }
 
 func serveHandler(cmd *cobra.Command, args []string) error {
-	go Container.Serve()
+	go func() {
+		Container.Serve()
+		stopSignal <- nil
+	}()
+
 	<-stopSignal
 	Container.Stop()