+++ cmd/rr/debug/debugger.go
@@ -5,6 +5,7 @@ import (
 	"github.com/spiral/roadrunner"
 	"github.com/spiral/roadrunner/cmd/rr/utils"
 	"github.com/spiral/roadrunner/service/http"
+	"strings"
 )
 
 // Listener creates new debug listener.
@@ -45,7 +46,6 @@ func (s *debugger) listener(event int, ctx interface{}) {
 			"<white+hb>worker.%v</reset> <yellow>killed</red>",
 			*w.Pid,
 		))
-
 	case roadrunner.EventWorkerError:
 		err := ctx.(roadrunner.WorkerError)
 		s.logger.Error(utils.Sprintf(
@@ -53,6 +53,12 @@ func (s *debugger) listener(event int, ctx interface{}) {
 			*err.Worker.Pid,
 			err.Caused,
 		))
+	}
+
+	// outputs
+	switch event {
+	case roadrunner.EventStderrOutput:
+		s.logger.Warning(strings.Trim(string(ctx.([]byte)), "\r\n"))
 	}
 
 	// rr server events
@@ -68,6 +74,8 @@ func (s *debugger) listener(event int, ctx interface{}) {
 	case roadrunner.EventPoolError:
 		s.logger.Error(utils.Sprintf("<red>%s</reset>", ctx))
 	}
+
+	//s.logger.Warning(event, ctx)
 }
 
 func statusColor(status int) string {